separate Interface logic and Serial logic
